+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - CLI flag `--env-file` added to the `blobl` command. (@mihaitodor)
+- Go API: New `ComponentLinter` APIs for linting general component configs. (@Jeffail)
 
 ### Fixed